TwoChoicesExtendedObliviousTransferProvider now lives in its own file.
