+++ CompactMPC/ObliviousTransfer/TwoChoicesExtendedOblivousTransferChannel.cs
@@ -85,23 +85,4 @@ namespace CompactMPC.ObliviousTransfer
             await CommunicationTools.WriteOptionsAsync(Channel, maskedOptions, 2, numberOfInvocations, numberOfMessageBytes);
         }
     }
-
-    public class TwoChoicesExtendedObliviousTransferProvider : ITwoChoicesObliviousTransferProvider
-    {
-        private ITwoChoicesObliviousTransferProvider _baseOT;
-        private int _securityParameter;
-        private CryptoContext _cryptoContext;
-
-        public TwoChoicesExtendedObliviousTransferProvider(ITwoChoicesObliviousTransferProvider baseOT, int securityParameter, CryptoContext cryptoContext)
-        {
-            _baseOT = baseOT;
-            _securityParameter = securityParameter;
-            _cryptoContext = cryptoContext;
-        }
-
-        public ITwoChoicesObliviousTransferChannel CreateChannel(IMessageChannel channel)
-        {
-            return new TwoChoicesExtendedObliviousTransferChannel(_baseOT.CreateChannel(channel), _securityParameter, _cryptoContext);
-        }
-    }
 }
